+++ apps/extension/src/app/query/bitcoin/balance/btc-balance.hooks.ts
@@ -2,6 +2,7 @@ import type { AccountId } from '@leather.io/models';
 import type { AccountRequest } from '@leather.io/services';
 import { createBtcBalance, createMoney } from '@leather.io/utils';
 
+import { useFlags } from '@app/features/feature-flags';
 import { useAccountAddresses } from '@app/services/accounts/use-account-addresses';
 import { useAccountRequest } from '@app/services/accounts/use-account-request';
 import { toFetchState } from '@app/services/fetch-state';
@@ -29,11 +30,14 @@ export function useCurrentBtcBalanceWithFallback() {
 export function useBtcAccountBalance(accountId: AccountId) {
   const account = useAccountAddresses(accountId);
   const discardedInscriptions = useDiscardedInscriptions();
+  const { isOrdinalsActive, isRunesActive } = useFlags();
   return toFetchState(
     useGetBtcAccountBalanceQuery({
       account,
       protections: {
         discardedInscriptions,
+        isOrdinalsActive,
+        isRunesActive,
       },
     })
   );
@@ -42,11 +46,15 @@ export function useBtcAccountBalance(accountId: AccountId) {
 export function useNativeSegwitBtcAccountBalance(accountId: AccountId) {
   const account = useAccountAddresses(accountId);
   const discardedInscriptions = useDiscardedInscriptions();
+  const { isOrdinalsActive, isRunesActive } = useFlags();
+
   return toFetchState(
     useGetBtcAccountBalanceQuery({
       account,
       protections: {
         discardedInscriptions,
+        isOrdinalsActive,
+        isRunesActive,
       },
       exclusions: { taprootAddresses: true },
     })
@@ -56,11 +64,15 @@ export function useNativeSegwitBtcAccountBalance(accountId: AccountId) {
 export function useTaprootBtcAccountBalance(accountId: AccountId) {
   const account = useAccountAddresses(accountId);
   const discardedInscriptions = useDiscardedInscriptions();
+  const { isOrdinalsActive, isRunesActive } = useFlags();
+
   return toFetchState(
     useGetBtcAccountBalanceQuery({
       account,
       protections: {
         discardedInscriptions,
+        isOrdinalsActive,
+        isRunesActive,
       },
       exclusions: { nativeSegwitAddresses: true },
     })