+++ apps/extension/src/app/query/collectibles/account-collectibles.query.ts
@@ -4,6 +4,7 @@ import { type CollectibleView, createCollectibleView } from '@leather.io/feature
 import { type AccountAddresses, type NonFungibleCryptoAsset } from '@leather.io/models';
 import { createAccountCollectiblesQueryConfig } from '@leather.io/queries';
 
+import { useFlags } from '@app/features/feature-flags';
 import { useUserSettings } from '@app/hooks/use-user-settings';
 
 function useAccountCollectiblesQuery(
@@ -11,10 +12,14 @@ function useAccountCollectiblesQuery(
   options: Partial<UseQueryOptions<NonFungibleCryptoAsset[], Error, CollectibleView[]>> = {}
 ) {
   const settings = useUserSettings();
+  const { isOrdinalsActive, isRunesActive } = useFlags();
   const { select, ...rest } = options;
 
   return useQuery<NonFungibleCryptoAsset[], Error, CollectibleView[]>({
-    ...createAccountCollectiblesQueryConfig({ account }, settings),
+    ...createAccountCollectiblesQueryConfig(
+      { account, protections: { isOrdinalsActive, isRunesActive } },
+      settings
+    ),
     ...rest,
     select: select ?? (collectibles => collectibles.map(createCollectibleView)),
   });