+++ packages/services/src/balances/runes-balances.service.ts
@@ -115,9 +115,11 @@ export class RunesBalancesService {
       const [taprootRunesOutputs, nativeSegwitRunesOutputs] = await Promise.all([
         this.bisApiClient.fetchRunesValidOutputs(request.account.bitcoin.taprootDescriptor, {
           signal,
+          isRunesActive: request.protections?.isRunesActive,
         }),
         this.bisApiClient.fetchRunesValidOutputs(request.account.bitcoin.nativeSegwitDescriptor, {
           signal,
+          isRunesActive: request.protections?.isRunesActive,
         }),
       ]);
       runesOutputs.push(...nativeSegwitRunesOutputs, ...taprootRunesOutputs);