+++ packages/ui/CHANGELOG.md
@@ -489,6 +489,15 @@
   * devDependencies
     * @leather.io/models bumped to 0.41.0
 
+## [1.111.3](https://github.com/leather-io/mono/compare/@leather.io/ui-v1.111.2...@leather.io/ui-v1.111.3) (2026-04-16)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @leather.io/features bumped to 1.4.2
+
 ## [1.111.2](https://github.com/leather-io/mono/compare/@leather.io/ui-v1.111.1...@leather.io/ui-v1.111.2) (2026-04-13)